refactor(shell): Phase 3c——OCR 采集与浮层生命周期外置（OCR 簇收官）
ocrCapture.js（325 行，15 个方法，5 段连续区间）：桌面走主进程 OverlayWindow、
浏览器走 getDisplayMedia 的采集入口；浮层开关、ESC 热键、document 级
down/move/up 框选监听与解绑。

project-overview.vue 4675 → 4357 行。累计 10639 → 4357（-59%）。
OCR 簇（33 个方法约 890 行）至此拆完。

这是风险最高的一组：持有跨方法的实例态（_ocrDoc* / _ocrGlobalBound /
_ocrKeydownBound），且 app-e2e 无法驱动。除常规验证外补了针对性检查：

- 监听器对称：模块内 addEventListener 的 7 类事件（mousedown/mousemove/
  mouseup/touchstart/touchmove/touchend/keydown）在同模块内均有对应
  removeEventListener，无跨模块悬挂
- 实例态不跨界：_ocrGlobalBound / _ocrKeydownBound / _ocrDocDown /
  _ocrDocMove / _ocrDocUp / _ocrKeydownHandler 六个字段在 .vue 与其他
  模块中引用数均为 0，全部收敛在本模块内
- 条件编译反向验证：本模块含 14 处 #ifdef（九个模块里最多），
  搬移后 #ifdef H5 内的文案在 h5 产物中仍在，未被误剥离

顺带修掉提取脚本的一个真缺陷：原先方法区间的 end 取"下一个方法起始行-1"，
会把前几期留下的「已外置」标记注释一并吞进新模块，同时从 .vue 里删掉——
而逐行比对**发现不了**（HEAD 在同位置也有该注释）。现改为把 end 锚定到方法
自身的闭合花括号，并加一道"区间内不得含标记行"的硬拦截。已复查前八个模块
均未中招（各自 grep 标记数为 0）。

验证：
- 逐行比对：5 段区间共 325 行字节一致
- 方法名扫描：九个模块 95 个方法与 .vue 剩余 119 个方法零重名
- npm run build:h5 通过；npm run check:emits 通过
- npm run test:app-e2e 3 次 2 通过 1 失败——失败是 J1 登录既有抖动
  （issue #200，master 同样 2/3），已用独立 puppeteer 探针确认本分支登录正常
